[ci] Set demand for macOS Mojave build machines
Context: https://github.com/xamarin/xamarin-android/pull/5589

We are working on upgrading our macOS build machines to macOS Catalina,
and in certain cases we'd like to continue to use an existing agent pool
rather than creating and authorizing a new one.  In those cases, the
existing pool will be updated to contain both Mojave and Catalina
machines.  We can make sure to always pick a Mojave machine by setting
a demand for the relevant system [capability][0] that any given
[machine may have][1].

[0]: https://docs.microsoft.com/en-us/azure/devops/pipelines/agents/agents?view=azure-devops&tabs=browser#capabilities
[1]: https://devdiv.visualstudio.com/DevDiv/_settings/agentqueues?agentId=688191&queueId=2206&view=capabilities