[decompress-assemblies] Add tool to decompress assemblies (#5536)
Commit d236af54 causes assemblies within `.apk` files to be
*separately* compressed.  Consequently, if you try to extract an
assembly from a `.apk` and read it, it fails:

	% cd samples/HelloWorld'
	% msbuild /t:SignAndroidPackage /p:Configuration=Release HelloWorld.csproj
	% unzip bin/Release/com.xamarin.android.helloworld-Signed.apk assemblies/HelloWorld.dll
	% ikdasm assemblies/HelloWorld.dll 

	Unhandled Exception:
	IKVM.Reflection.BadImageFormatException: Exception of type 'IKVM.Reflection.BadImageFormatException' was thrown.

	% file assemblies/HelloWorld.dll 
	assemblies/HelloWorld.dll: Sony PlayStation Audio

(The `file` output is rather humorous…)

The `assemblies/*.dll` files still contain IL, they're just LZ4
compressed following a `CompressedAssemblyHeader`.

A way to read the assemblies is to instead use the build tree,
using e.g. `obj/Release/android/assets/HelloWorld.dll`.

Instead of requiring the build tree, add a new `decompress-assemblies`
utility which is able to uncompress the assemblies.

Usage:

	decompress-assemblies FILES*

where `FILES` are `.dll` files, e.g. as extracted via `unzip` (above),
`.apk` files, or `.aab` files.  If `.apk` or `.aab` files are provided,
then all `assemblies/*.dll` files within the `.apk` or `.aab` will be
extracted and uncompressed.

The uncompressed result is stored in the current directory with the
original assembly name modified to contain the `uncompressed-` prefix,
in order to prevent accidental overwriting of assemblies found in the
current directory.