[ci] Move TimeZoneInfo tests to Nightly pipeline (#5568)
The broad set of TimeZoneInfo deployment tests don't need to be running
against every commit and PR build.  They should be moved to the Nightly
build and test pipeline that was introduced in commit 916d24b.  This 
will reduce PR build execution times and the overall machine time cost
of PR/CI builds.