[Xamarin.Android.Build.Tasks] bump to System.Collections.Immutable 1.7.1
I keep getting this error in VS Windows:

    Detected package downgrade: System.Collections.Immutable from 1.7.1 to 1.7.0. Reference the package directly from the project to select a different version.
        Xamarin.Android.Build.Tasks -> Microsoft.Android.Build.BaseTasks -> System.Collections.Immutable (>= 1.7.1)
        Xamarin.Android.Build.Tasks -> System.Collections.Immutable (>= 1.7.0)

But command-line it seems to work fine?

It does seem like this package is behind compared to one used in
`xamarin-android-tools`, so let's just update it.

I also went ahead and created `$(SystemCollectionsImmutableVersion)`
in `Directory.Build.props` as this NuGet package is used in a few
places.