[jnienv-gen] Generate a header file for the native functions (#809)
Update `jnienv-gen` so that in addition to generate `.c` and `.cs`
files, it also generates a `.h` file with declarations for the
functions in the `.c` file.

The header will be used by the NET6 build of Xamarin.Android's
native runtime.

Additionally, make it possible to *not* export symbols by defining
the `JI_NO_VISIBILITY` preprocessor macro.  This will also be used by
the NET6 build of Xamarin.Android native runtime (which will no
longer need to make the symbols visible for `p/invoke` to work).