[MSBuildDeviceIntegration] Give each test a unique package name (#5551)
One of the problems we occasionally see on CI is when two tests run
at the same time on the same emulator.  This can cause collisions and
cause tests to fail because the app was changed while the test was
running.

Currently ALL of the tests user `UnnamedProject.UnnamedProject` as a
package name.  This means they overwrite each other.

Introduce a new system which will auto generate a package name based
on the test name itself.  We do this by using the
[`[CallerMemberNameAttribute]`][0]  attribute to determine the name
of the method which called the constructor of
`XamarinAndroidApplicationProject`.  This is then used to construct
a unique (ish) package name in the form `com.xamarin.<TestMethodName>`.

This should reduce collisions.  As part of the process we also
register each `packageName` used with a static cache.  We then
uninstall those from the device when the `[OneTimeTearDown]` method
is called.  This means the device will be cleaned as groups of tests
are run.

[0]: https://docs.microsoft.com/dotnet/api/system.runtime.compilerservices.callermembernameattribute