[Xamarin.Android.Build.Tasks] Add $(AndroidLinkResources) (#5317)
The generated `Resource` class in `Resource.designer.cs` can be quite
large.

When doing "Release" builds, update the linker to *remove* the
`Resource` class and its nested types.  We do this by examining all
the locations where the fields are used and replacing those IL calls
with a call which specifies the actual resource id value directly.

This will remove the need to call
`global::Android.Runtime.ResourceIdManager.UpdateIdValues();` from
the static constructors in the `Resource` class and its nested types,
which will improve app launch times (as `UpdateIdValues()` involves
Reflection, which is not "fast").

Due to an odd linker failure we cannot completely remove the
`Resource` class itself.  It *will* be completely empty.

"Debug" builds will *not* remove the `Resource` type, and will retain
the call to `UpdateIdValues()`, in order to ensure that app rebuild
and redeploy times are as fast as possible.

The new `$(AndroidLinkResources)` MSBuild property has been added to
control this feature.  It is disabled by default for now.

`$(AndroidLinkResources)`=True appears to save between 100-280kb for
an `.apk`.  This is dependent on how many assemblies are Android
assemblies and contain or use Resources.

Xamarin.Forms base app `apkdiff` results

	Size difference in bytes ([*1] apk1 only, [*2] apk2 only):
	  -          10 assemblies/Mono.Android.dll
	  -       4,001 assemblies/Xamarin.Essentials.dll
	  -      45,426 assemblies/Xamarin.Forms.Platform.Android.dll
	  -     102,839 assemblies/formstest.Android.dll
	Summary:
	  +           0 Other entries 0.00% (of 887,808)
	  +           0 Dalvik executables 0.00% (of 3,341,152)
	  -     152,276 Assemblies -3.23% (of 4,709,077)
	  +           0 Shared libraries 0.00% (of 41,693,796)
	  -     151,552 Package size difference -0.73% (of 20,820,695)

Maps sample `apkdiff` results

	Size difference in bytes ([*1] apk1 only, [*2] apk2 only):
	  -           4 lib/armeabi-v7a/libxamarin-app.so
	  -          14 assemblies/Mono.Android.dll
	  -       4,757 assemblies/Xamarin.Essentials.dll
	  -      46,051 assemblies/Xamarin.Forms.Platform.dll
	  -      47,586 assemblies/Xamarin.Forms.Platform.Android.dll
	  -      48,470 assemblies/Xamarin.Forms.Maps.Android.dll
	  -     135,589 assemblies/FormsMapsSample.Android.dll
	Summary:
	  +           0 Other entries 0.00% (of 1,137,209)
	  +           0 Dalvik executables 0.00% (of 2,519,972)
	  -     282,467 Assemblies -5.51% (of 5,123,833)
	  -           4 Shared libraries -0.00% (of 41,843,596)
	  -     282,624 Package size difference -1.34% (of 21,075,664)