Output Xamarin.Android and MonoVM version information in logcat
With the upcoming NET6 release, both Xamarin.Android itself and
Xamarin.Android applications will be built against a set of Mono VM
runtimes installed from nuget packages.  This may create a situation
when the Mono VM runtime nuget is updated by Xamarin.Android is not - it
will then run on the device with a different runtime version that it was
built against.

Ideally, the above should not create any issues, but occasionally it may
happen that Mono VM behavior changes and causes an issue in
Xamarin.Android that leads to a crash.  In such instance it would be
useful what version of Xamarin.Android the application was running and
in what version of MonoVM it was executing.

This PR logs the version information to logcat, just before handing over
control to the managed code.