[monodroid] Print version information to logcat (#5677)
With the upcoming .NET 6 release, both Xamarin.Android itself and
Xamarin.Android applications will be built against a set of MonoVM
runtimes installed via NuGet packages.  This may create a situation
where the MonoVM runtime NuGet is updated but Xamarin.Android is not;
it will then run on the device with a different runtime version that
it was built against.

Ideally, the above should not create any issues, but occasionally it
may happen that MonoVM or .NET 6 BCL behavior changes and causes an
issue in Xamarin.Android that leads to a crash.  In such instances it
would be useful to know what version of Xamarin.Android the
application was running against and what the MonoVM version info is.

Add `log_info()` calls to print out the Xamarin.Android version
information and the result of calling `mono_get_runtime_build_info()`
to `adb logcat`, just before handing over control to managed code.