[One .NET] AutoImport.props should include Transforms.xml (#5703)
Context: https://github.com/xamarin/XamarinComponents/tree/master/Android/Guava/source/Guava

Looking at some of the patterns in the XamarinComponents repo, many
of the bindings commonly have a single `Transforms.xml` in the root
of the project. I think this pattern makes complete sense, so let's
add this to `AutoImport.props` by default.

I updated a test for this scenario, the test relies on the
`@(TransformFile)` wildcards working.