[doc] More info about size regression checks (#5718)
Add more detail about apk size regression testing.

Use `apkdiff --descrease-is-regression` for `.apk` instrumentation
tests.

Mention `Documentation/project-docs/ApkSizeRegressionChecks.md`
in the failed check messages.

Update the `.apkdesc` reference files.  The change with
`apkdiff --descrease-is-regression` kicked in and we need to refresh
the reference files, because we have smaller `.apk` files compared to
the last time we updated these files.