[Mono.Android] Update BuildVersionCodes.R value to 30. (#5725)
Fixes: https://github.com/xamarin/xamarin-android/issues/5723

Our enumification process stores the value of the constants in
`map.csv`.  The build version code constant for a preview API level
is 10000, but changes to e.g.: 30 when the API goes stable.

Depending on when enumification was done, the 10000 may be
accidentally stored instead of 30.

When the API goes stable we must update `map.csv` to have the
correct value.

Update the enum value `Android.OS.BuildVersionCodes.R` to `30`.

Add this note to the in-progress "Adding a new API Level" checklist
for our future benefit.