[CI] Add Linux Build and Smoke Test job (#5681)
A new Linux Build + Smoke test job has been added to produce a
"commercial" .NET 6 Linux SDK `.nupkg`.  This job takes advantage of
a new [Azure Virtual Machine Scale Set (VMSS)][0] pool that has been
created specifically for this job.  This VMSS can be found in the
Azure portal under the name [xaubuntu20.04-vmss][1].

This job is configured to run the `PackagingTest.CheckSignApk()`\
tests, a subset of the broader smoke test category.  More tests should
be added to this job in the future as needed.  The execution of these
tests on this new build pool exposed a test framework issue that occurs
on environments that do not export `$JAVA_HOME`, or have a JDK
installation in any default look up paths:

	Failed CheckSignApk(True,False) [27 ms]
	  Error Message:
	   System.InvalidOperationException : Could not determine Java SDK location. Please provide `javaSdkPath`.
	  Stack Trace:
	     at Xamarin.Android.Tools.AndroidSdkInfo..ctor(Action`2 logger, String androidSdkPath, String androidNdkPath, String javaSdkPath) in /agent/_work/1/s/xamarin-android/external/xamarin-android-tools/src/Xamarin.Android.Tools.AndroidSdk/AndroidSdkInfo.cs:line 25
	   at Xamarin.Android.Build.Tests.PackagingTest.CheckSignApk(Boolean useApkSigner, Boolean perAbiApk) in /agent/_work/1/s/xamarin-android/src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/PackagingTest.cs:line 366

To avoid this, we can pass the JDK path used by the tests when
creating new instances of `AndroidSdkInfo`.  In some of these cases
`AndroidSdkInfo` was only being used to calculate a JDK path for a
tool that a test needed, and in these cases we can use
`AndroidSdkResolver.GetJavaSdkPath()` directly.

[0]: https://docs.microsoft.com/en-us/azure/devops/pipelines/agents/scale-set-agents?view=azure-devops
[1]: https://ms.portal.azure.com/#@microsoft.onmicrosoft.com/resource/subscriptions/cd4829e2-e38b-43d2-8316-2f2009f36f97/resourceGroups/DEVDIV-VM-SCALE-SETS/providers/Microsoft.Compute/virtualMachineScaleSets/xaubuntu20.04-vmss/overview