[Xamarin.Android.Build.Tasks] make a way to keep LaunchProfile capability (#5745)
Context: https://github.com/microsoft/VSProjectSystem/blob/master/doc/overview/about_project_capabilities.md

A new `$(_KeepLaunchProfiles)` MSBuild property will be defined on
single projects to use `LaunchProfiles` capability required on
VisualStudio, see:

https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1291515

When a single-project is loaded in VisualStudio, to enable the UI for
changing between target frameworks the `LaunchProfiles` capability
needs to be enabled.

By default, this capability is removed by the iOS & Android SDKs
because the HEAD projects (Android and iOS) does not use this IDE
feature for displaying devices on the start/play button.

The new `$(_KeepLaunchProfiles)` property avoids that removal in order
to have `LaunchProfiles` enabled in VisualStudio. MAUI's MSBuild targets
will set `$(_KeepLaunchProfiles)` to enable it for single-project.

Eventually, we may be able to remove `$(_KeepLaunchProfiles)` completely
when Maui is a .NET workload. In that case, we will have better control
of MSBuild ordering. However, for the immediate .NET 6 Previews, Maui
will be shipping as a NuGet package.