OpenTK Deprecation
We are going to stop shipping OpenTK-1.0 in a future update. In order to give users who are using
that assembly a chance to migrate we will deprecate it first.
We will raise a `XA4313` warning if we detect that OpenTK-1.0 is in the `@(Refereces)` item group.

A replacement Nuget Package has already been created at https://www.nuget.org/packages/Xamarin.Legacy.OpenTK/
Users seeing this warning should remove the reference to `OpenTK-1.0` and replace it will the following

```
<PackageReference Include="Xamarin.Legacy.OpenTK" Version="0.0.1-alpha" />
```

The version number will change in the future.

The vast majority of changes in this commit are to stop the OpenTK java types being included in `mono-android.jar`.
From this point on `OpenTK` will be treated like all other Android Class Libraries, the generated java code will be
written to the `$(IntermediateOutputPath)\lp` directory along with all the other references.