[Xamarin.Android.Build.Tasks] Xamarin.Android deployment fails unless project is cleaned first
Fixes https://work.azdo.io/1288593

This is a weird bug which I cannot reproduce locally. However reading
the code this code path is possible.

When trying to do incremental builds when using the `AndroidPackageFormat`
`aab` we get the following error

     [BT : 1.4.0] error : Module 'base' is missing mandatory file 'manifest/AndroidManifest.xml'.

The problem is the base.zip doesn't have ANY AndroidManifest.xml in it...
But why?

Well what seems to happen is in this case the input file `packaged_resources`
and the output `base.zip` are unchanged. `packaged_resources` has
an `AndroidManifest.xml` file and `base.zip` has a `manifest/AndroidManifest.xml`
file. Because it was renamed from `AndroidManifest.xml` to `manifest/AndroidManifest.xml`
on a previous build. This is so that our `base.zip` confirms to the expected
folder structure for an `aab`. So this is by design.

However, in this case because BOTH files are up to date, we DONT add
the `AndroidManifest.xml` file from `packaged_resources` to the `base.zip`.
A compounding problem is because no physical file `manifest/AndroidManifest.xml`
exists, that file is then removed from the `base.zip`.
As a result the `FixupArchive` which does the renaming of the `AndroidManifest.xml`
for `aab` files has nothing to do. Because its source file does not exist.

So we end up with a `base.zip` without ANY `AndroidManifest.xml` file in it.

The fix in this case is to NEVER remove an `AndroidManifest.xml` file from
the base.zip or apk. It is a protected file and should always exist.