[build] Include Configuration.OperatingSystem.props earlier
This allows the other files (e.g. `Configuration.Override.props`) to
rely on the properties (e.g. `$(HostOS)`) defined in
`Configuration.OperatingSystem.props`. This is helpful when sharing the
override file between various operating systems as it allows one to
specify the ABIS to build in a straightforward manner, e.g.:

    <!-- This is Configuration.Override.props -->
    <Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
      <PropertyGroup>
        <AndroidSupportedHostJitAbis>$(HostOS):mxe-Win32:mxe-Win64</AndroidSupportedHostJitAbis>
      </PropertyGroup>
    </Project>