[Xamarin.Android.Build.Tasks] jarsigner file permission warning
We are encountering some warnings when running `jarsigner` from
Microsoft OpenJDK 11:

    C:\Users\AzDevOps\android-toolchain\jdk-11\bin\jarsigner.exe : warning ANDJS0000: POSIX file permission and/or symlink attributes detected. These attributes are ignored when signing and are not protected by the signature.
    C:\Users\AzDevOps\android-toolchain\jdk-11\bin\jarsigner.exe : warning ANDJS0000: POSIX file permission and/or symlink attributes detected. These attributes are ignored when signing and are not protected by the signature.

We believe the problem lies with libZipSharp setting a default Unix
permission on Windows:

https://github.com/xamarin/LibZipSharp/blob/05f94a482fe9a791691d22951c3345ee67827b54/WindowsPlatformServices.cs#L199

Since, we're unsure why this was done, let's downgrade the warning to a
message instead. We already had an `IgnoredWarnings` list we could add
the message to in the `<AndroidSignPackage/>` MSBuild task.