[azdo] restore with MSBuild instead of NuGet (#5805)
This should allow us to get `.binlog` files for diagnosing when things
go wrong. `nuget restore -Verbosity detailed` is not as detailed as we
would like.

Use `$(RestoreConfigFile)`, which maps to the `-ConfigFile` switch for
NuGet:

https://docs.microsoft.com/nuget/reference/msbuild-targets

We no longer need to run the `NuGetToolInstaller` yaml task either.

Down the road, we can hopefully switch to using `dotnet restore`. But
we can't do this yet, because `Xamarin.Android-Tests.sln` contains "legacy"
Xamarin.Android projects.