[Xamarin.Android.Build.Tasks] Add $(UseInterpreter) (#5839)
Fixes: https://github.com/xamarin/xamarin-android/issues/5824
Context: https://github.com/mono/mono/issues/19669

A new `$(UseInterpreter)` property has been added to toggle the existing
`$(AndroidUseInterpreter)` setting.  This new property does not contain
OS / Platform information so that it can be used by multi-targeted
projects.

The `<BuildApk/>` task no longer needs to check the list of supported
ABIs when the interpreter is enabled.  The interpreter should now work
with all architectures.  I ran the .NET 6 version of the new test on an
x86 emulator locally and the app launched successfully.

There was also an oversight in the legacy targets that caused
interpreter builds to fail with the following, which has been fixed:

    (_BuildApkEmbed target) ->
      C:\Program Files (x86)\Microsoft Visual Studio\2019\Enterprise\MSBuild\Xamarin\Android\Xamarin.Android.Common.targets(2010,3): error XABLD7028: System.IO.FileNotFoundException: Could not find file 'C:\Program Files (x86)\Microsoft Visual Studio\2019\Enterprise\MSBuild\Xamarin\Android\lib\interpreter-arm64-v8a\libxa-internal-api.so'.
      C:\Program Files (x86)\Microsoft Visual Studio\2019\Enterprise\MSBuild\Xamarin\Android\Xamarin.Android.Common.targets(2010,3): error XABLD7028: File name: 'C:\Program Files (x86)\Microsoft Visual Studio\2019\Enterprise\MSBuild\Xamarin\Android\lib\interpreter-arm64-v8a\libxa-internal-api.so'