[tests] fixed CheckTimestamps
The path to the linker directory was expecting:

    obj\Debug\net6.0-android\linked

When we are now using:

    obj\Debug\net6.0-android\[RID]\linked

I updated the test to just check for a parent directory named `linked`.

This was passing in past .NET 6 builds, but looking a the MSBuild
targets in dotnet/sdk, it does not seem like they even care about the
timestamps on linker output:

https://github.com/dotnet/sdk/blob/a245b6ff06b483927e57d953b803a390ad31db95/src/Tasks/Microsoft.NET.Build.Tasks/targets/Microsoft.NET.ILLink.targets#L113-L116

`Outputs="$(_LinkSemaphore)"` will likely be the only file that gets
"touch"-ed.