[ci] Use dotnet to create workload .pkg (#5842)
Commit 12edc2c5 introduced a feature that seems to only be available
in newer versions of MSBuild:

    <Import Project="Sdk.props" Sdk="Microsoft.DotNet.SharedFramework.Sdk" Version="$(MicrosoftDotNetBuildTasksFeedPackageVersion)" />

This caused an issue during the `dotnet_create_pkg` job on our
Xamarin.Android pipeline, which uses a different machine pool than
Xamarin.Android-PR:

    /Users/builder/azdo/_work/2/s/xamarin-android/build-tools/create-packs/Directory.Build.targets(11,3): error : Version 5.0.201 of the .NET Core SDK requires at least version 16.8.0 of MSBuild. The current available version of MSBuild is 16.6.0.52201. Change the .NET Core SDK specified in global.json to an older version that requires the MSBuild version currently available.

    "/Users/builder/azdo/_work/2/s/xamarin-android/Xamarin.Android.sln" (CreateWorkloadInstallers target) (1) ->
    "/Users/builder/azdo/_work/2/s/xamarin-android/build-tools/create-packs/Microsoft.Android.Sdk.proj" (ExtractWorkloadPacks target) (2) ->
      /Users/builder/azdo/_work/2/s/xamarin-android/build-tools/create-packs/Directory.Build.targets(11,31): error MSB4236: The SDK 'Microsoft.DotNet.SharedFramework.Sdk/$(MicrosoftDotNetBuildTasksFeedPackageVersion)' specified could not be found. [/Users/builder/azdo/_work/2/s/xamarin-android/build-tools/create-packs/Microsoft.Android.Sdk.proj]

I believe this wasn't seen on the PR bots as some of them now have the
a newer mono installed which supports MSBuild 16.9 (see PR #5806, which
is still waiting for designer test fixes).

Fix this issue by using `dotnet build` on macOS when we build the
`CreateWorkloadInstallers` target.

Finally, I hit a considerable amount of disk space related issues when
testing this PR which will hopefully be alleviated by migrating to the
MicroBuild Virtual Machine Scale Set pool.  I've also moved the Build
Asset Registry manifest promotion to the end of the job so that it won't
run multiple times when retrying a failed job.