[One .NET] exclude libxamarin-debug-app-helper.so from Release builds (#5874)
Fixes: https://github.com/xamarin/xamarin-android/issues/5855

Currently, in .NET 6, any `.so` file in our runtime pack is included
in apps by default. This is the nature of how native libraries coming
from NuGet packages work. They just get added based on the file path
and `$(RuntimeIdentifier)` of the project.

We already have a special case for `libmono-android.release.so` and
`libmono-android.debug.so` that uses `$(AndroidIncludeDebugSymbols)`
as a way to know which native library is needed.

We can do the same thing here and exclude
`libxamarin-debug-app-helper.so`, which should only be included in
`Debug` builds.

This saves ~37K for all Release apps:

    -37,104 lib/arm64-v8a/libxamarin-debug-app-helper.so *1